new parser working for previous assignment

+++ A6-Resources/check_spim.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 clear
 
+# usage --> ./check_spim 0 --> run without icode
+# usage --> ./check_spim 1 --> run with icode
+
+
 pass=0
 fail=0
 pass2=0
@@ -11,43 +15,56 @@ pass4=0
 fail4=0
 
 
-for i in $(find TestCases -name "test2.c");
+for i in $(find TestCases -name "test.c");
 do 
-    ./sclp_old -ast -symtab -icode $i # sclp_old
+    if [ "$1" == "1" ]
+    then
+        ./sclp_old -ast -symtab -icode $i # sclp_old
+        cp $i.ic test_icode_old.txt
+    else
+        ./sclp_old -ast -symtab $i # sclp_old
+    fi
+
     cp $i.ast test_ast_old.txt
     cp $i.spim test_spim_old.txt
     cp $i.sym test_symtab_old.txt
-    cp $i.ic test_icode_old.txt
 
-    ./sclp_new -ast -symtab -icode $i # sclp_new
+#----------------------------------------#
+    if [ "$1" == "1" ]
+    then
+        ./sclp_new -ast -symtab -icode $i # sclp_new
+        cp $i.ic test_icode_new.txt
+    else
+        ./sclp_new -ast -symtab $i # sclp_new
+    fi
+
     cp $i.ast test_ast_new.txt
     cp $i.spim test_spim_new.txt
     cp $i.sym test_symtab_new.txt
-    cp $i.ic test_icode_new.txt
 
     #--------test_ast----------
     
     DIFF=$(diff test_ast_old.txt test_ast_new.txt) 
     if [ "$DIFF" == "" ] 
     then
-    pass=$((pass+1))
-     echo "Passed $i ast"
+        pass=$((pass+1))
+        echo "Passed $i ast"
    
     else
-    fail=$((fail+1))
-    echo "Failed $i ast"
+        fail=$((fail+1))
+        echo "Failed $i ast <<======="
     fi
 
     #-----------test_spim---------------
     DIFF1=$(diff test_spim_old.txt test_spim_new.txt) 
     if [ "$DIFF1" == "" ] 
     then
-    pass2=$((pass2+1))
-     echo "Passed $i spim"
+        pass2=$((pass2+1))
+        echo "Passed $i spim"
     
     else
-    fail2=$((fail2+1))
-    echo "Failed $i spim"
+        fail2=$((fail2+1))
+        echo "Failed $i spim <<======="
     fi
 
     #----------------test_symtab------------
@@ -55,25 +72,28 @@ do
     DIFF2=$(diff test_symtab_old.txt test_symtab_new.txt) 
     if [ "$DIFF2" == "" ] 
     then
-    pass3=$((pass3+1))
-     echo "Passed $i symtab"
+        pass3=$((pass3+1))
+        echo "Passed $i symtab"
     
     else
-    fail3=$((fail3+1))
-    echo "Failed $i symtab"
+        fail3=$((fail3+1))
+        echo "Failed $i symtab <<======="
     fi
 
     #----------------test_icode---------------
 
-    DIFF2=$(diff test_icode_old.txt test_icode_new.txt) 
-    if [ "$DIFF2" == "" ] 
+    if [ "$1" == "1" ]
     then
-    pass4=$((pass4+1))
-     echo "Passed $i icode"
-    
-    else
-    fail4=$((fail4+1))
-    echo "Failed $i icode"
+        DIFF2=$(diff test_icode_old.txt test_icode_new.txt) 
+        if [ "$DIFF2" == "" ] 
+        then
+            pass4=$((pass4+1))
+            echo "Passed $i icode"
+        
+        else
+            fail4=$((fail4+1))
+            echo "Failed $i icode <<======="
+        fi
     fi
 
     echo -e "-------------------\n\n"
@@ -81,7 +101,11 @@ do
     rm $i.ast
     rm $i.spim
     rm $i.sym
-    rm $i.ic
+
+    if [ "$1" == "1" ]
+    then
+        rm $i.ic
+    fi
 
 done
 #rm test_ast_old.txt
@@ -96,5 +120,8 @@ echo -e "Failed Cases $fail2 spim\n"
 echo -e "Passed Cases $pass3 symtab"
 echo -e "Failed Cases $fail3 symtab\n"
 
-echo -e "Passed Cases $pass4 icode"
-echo -e "Failed Cases $fail4 icode\n"
+if [ "$1" == "1" ]
+    then
+    echo -e "Passed Cases $pass4 icode"
+    echo -e "Failed Cases $fail4 icode\n"
+fi
